+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming...
+
+- Examples of handling errors
+
 ## 0.1.0 (2020-09-02)
 
 - First minimal examples